refactor: move telegram utils
