+++ src/lib/telegram/utils.ts
@@ -1,3 +1,5 @@
+import { GrammyError, type Context } from "grammy/web";
+
 export function normalizeUserMention({
   text,
   username,
@@ -14,4 +16,16 @@ export function normalizeUserMention({
     }
   }
   return text;
+}
+
+export function formatTelegramSessionName(context: Context): string {
+  return ["tg", context.chat?.id ?? "unknown", context.message?.message_thread_id]
+    .filter(Boolean)
+    .join("-");
+}
+
+export function getTelegramRetryAfter(error: unknown): number | undefined {
+  if (error instanceof GrammyError && error.error_code === 429) {
+    return error.parameters.retry_after;
+  }
 }
